add HistoryVerifier
