+++ packages/plasma-light-client/src/verifier/index.ts
@@ -1,0 +1,1 @@
+export * from './HistoryVerifier'
